Less hacky KTX2Loader and BasisTextureLoader
